Use std::multiset as the event queue
This allows us to delete nodes in batch, at the moment we delete
existing ones.

Also we tweak the allocators, using a pool allocator for the set and the
interval_map data structures. With this, it makes no sense to have a
custom allocator in the interface of the split_rectangles function
anymore.